实现能量生成炮塔

+++ Assets/GameMain/Scripts/Item/ItemTowerBuildButton.cs
@@ -78,7 +78,7 @@ namespace Flower
             this.onClick = onClick;
         }
 
-        private void UpdateEnergyState(int ownEnergy)
+        private void UpdateEnergyState(float ownEnergy)
         {
             if (towerLevelData == null)
                 return;
